+++ internal/repository/repository.go
@@ -36,6 +36,33 @@ func Use(ID int) interface{} {
 }
 
 // All returns all objects currently stored.
-func All() []interface{} {
-	return storedObjects
+func All(from, limit int) (map[int]interface{}, bool) {
+	if from < len(storedObjects) {
+		if from+limit < len(storedObjects) {
+			return tomap(from, from+limit), false
+		}
+		return tomap(from, len(storedObjects)), from == 0
+	}
+	if from > len(storedObjects) {
+		return nil, false
+	}
+	return map[int]interface{}{}, from == 0
+}
+
+// Reset removes all objects in storage.
+func Reset() {
+	storedObjects = []interface{}{}
+}
+
+// Len gives the total number of objects stored.
+func Len() int {
+	return len(storedObjects)
+}
+
+func tomap(start, end int) map[int]interface{} {
+	m := make(map[int]interface{})
+	for i := start; i < end; i++ {
+		m[i] = storedObjects[i]
+	}
+	return m
 }