+++ pkg/smtpd/session_test.go
@@ -242,10 +242,10 @@ func test(t *testing.T, snd string, rcv string) {
 	c := textproto.NewConn(rwc)
 	assert.NotNil(t, c, "")
 
-	s := smtpd.NewSession(c, nil)
+	s := smtpd.NewSession(c, nil, nil)
 	assert.NotNil(t, s, "")
 
-	s.Serve()
+	s.Serve(make(chan struct{}, 1))
 
 	responses, err := ioutil.ReadAll(rcvbuf)
 	assert.NoError(t, err, "")